// dict/api/appointment/types.ts
/* eslint-disable */
// tslint:disable
/**
 * Patients Appointment
 * Patient Appointment
 *
 * OpenAPI spec version: v1
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator+.
 * https://github.com/karlvr/openapi-generator-plus
 * Do not edit the class manually.
 */

import { Express } from 'express'
import { Api } from '../../models'

export interface AppointmentApi {
	postAppointmentCreate: (request: Api.AppointmentDt | undefined) => Promise<PostAppointmentCreateResponse>
	deleteAppointmentDelete: (id: string) => Promise<DeleteAppointmentDeleteResponse>
	getAppointmentGet: (id: string) => Promise<GetAppointmentGetResponse>
	getAppointmentGetAll: (limit: number | null | undefined, direction: Api.DirectionParamEnum | undefined, sortByField: string | null | undefined) => Promise<GetAppointmentGetAllResponse>
	putAppointmentUpdate: (id: string | undefined, request: Api.AppointmentDt | undefined) => Promise<PutAppointmentUpdateResponse>
}

export type PostAppointmentCreateResponse = PostAppointmentCreate201Response | PostAppointmentCreate401Response | PostAppointmentCreate404Response | PostAppointmentCreate422Response

export interface PostAppointmentCreate201Response {
	status: 201
	body: Api.AppointmentDt
	headers?: never
}

export interface PostAppointmentCreate401Response {
	status: 401
	body: Api.MessageDto
	headers?: never
}

export interface PostAppointmentCreate404Response {
	status: 404
	body: Api.MessageDto
	headers?: never
}

export interface PostAppointmentCreate422Response {
	status: 422
	body: Api.MessageDto
	headers?: never
}

export type DeleteAppointmentDeleteResponse = DeleteAppointmentDelete200Response | DeleteAppointmentDelete401Response | DeleteAppointmentDelete404Response | DeleteAppointmentDelete422Response

export interface DeleteAppointmentDelete200Response {
	status: 200
	body: Api.MessageDto
	headers?: never
}

export interface DeleteAppointmentDelete401Response {
	status: 401
	body: Api.MessageDto
	headers?: never
}

export interface DeleteAppointmentDelete404Response {
	status: 404
	body: Api.MessageDto
	headers?: never
}

export interface DeleteAppointmentDelete422Response {
	status: 422
	body: Api.MessageDto
	headers?: never
}

export type GetAppointmentGetResponse = GetAppointmentGet200Response | GetAppointmentGet401Response | GetAppointmentGet404Response | GetAppointmentGet422Response

export interface GetAppointmentGet200Response {
	status: 200
	body: Api.AppointmentDt
	headers?: never
}

export interface GetAppointmentGet401Response {
	status: 401
	body: Api.MessageDto
	headers?: never
}

export interface GetAppointmentGet404Response {
	status: 404
	body: Api.MessageDto
	headers?: never
}

export interface GetAppointmentGet422Response {
	status: 422
	body: Api.MessageDto
	headers?: never
}

export type GetAppointmentGetAllResponse = GetAppointmentGetAll200Response | GetAppointmentGetAll401Response | GetAppointmentGetAll404Response

export interface GetAppointmentGetAll200Response {
	status: 200
	body: Api.AppointmentPagedResultDto
	headers?: never
}

export interface GetAppointmentGetAll401Response {
	status: 401
	body: Api.MessageDto
	headers?: never
}

export interface GetAppointmentGetAll404Response {
	status: 404
	body: Api.MessageDto
	headers?: never
}

export type PutAppointmentUpdateResponse = PutAppointmentUpdate200Response | PutAppointmentUpdate401Response | PutAppointmentUpdate404Response | PutAppointmentUpdate422Response

export interface PutAppointmentUpdate200Response {
	status: 200
	body: Api.AppointmentDt
	headers?: never
}

export interface PutAppointmentUpdate401Response {
	status: 401
	body: Api.MessageDto
	headers?: never
}

export interface PutAppointmentUpdate404Response {
	status: 404
	body: Api.MessageDto
	headers?: never
}

export interface PutAppointmentUpdate422Response {
	status: 422
	body: Api.MessageDto
	headers?: never
}